split fault from outcomes stratgy
