+++ src/Polly.Core/Simmy/Fault/FaultConstants.cs
@@ -1,0 +1,6 @@
+﻿namespace Polly.Simmy.Fault;
+
+internal static class FaultConstants
+{
+    public const string OnFaultInjectedEvent = "OnFaultInjectedEvent";
+}
