+++ src/Polly.Core/Simmy/Fault/FaultGeneratorArguments.cs
@@ -1,4 +1,4 @@
-﻿namespace Polly.Simmy.Outcomes;
+﻿namespace Polly.Simmy.Fault;
 
 #pragma warning disable CA1815 // Override equals and operator equals on value types
