+++ src/Polly.Core/Simmy/Fault/FaultPipelineBuilderExtensions.TResult.cs
@@ -1,22 +1,23 @@
 ﻿using System.Diagnostics.CodeAnalysis;
-using Polly.Simmy.Outcomes;
+using Polly.Simmy.Fault;
 
 namespace Polly.Simmy;
 
 /// <summary>
 /// Extension methods for adding outcome to a <see cref="ResiliencePipelineBuilder"/>.
 /// </summary>
-internal static partial class OutcomePipelineBuilderExtensions
+internal static partial class FaultPipelineBuilderExtensions
 {
     /// <summary>
     /// Adds a fault chaos strategy to the builder.
     /// </summary>
+    /// <typeparam name="TResult">The type of result the retry strategy handles.</typeparam>
     /// <param name="builder">The builder instance.</param>
     /// <param name="enabled">A value that indicates whether or not the chaos strategy is enabled for a given execution.</param>
     /// <param name="injectionRate">The injection rate for a given execution, which the value should be between [0, 1] (inclusive).</param>
     /// <param name="fault">The exception to inject.</param>
     /// <returns>The builder instance with the retry strategy added.</returns>
-    public static ResiliencePipelineBuilder AddChaosFault(this ResiliencePipelineBuilder builder, bool enabled, double injectionRate, Exception fault)
+    public static ResiliencePipelineBuilder<TResult> AddChaosFault<TResult>(this ResiliencePipelineBuilder<TResult> builder, bool enabled, double injectionRate, Exception fault)
     {
         Guard.NotNull(builder);
 
@@ -32,13 +33,14 @@ internal static partial class OutcomePipelineBuilderExtensions
     /// <summary>
     /// Adds a fault chaos strategy to the builder.
     /// </summary>
+    /// <typeparam name="TResult">The type of result the retry strategy handles.</typeparam>
     /// <param name="builder">The builder instance.</param>
     /// <param name="enabled">A value that indicates whether or not the chaos strategy is enabled for a given execution.</param>
     /// <param name="injectionRate">The injection rate for a given execution, which the value should be between [0, 1] (inclusive).</param>
     /// <param name="faultGenerator">The exception generator delegate.</param>
     /// <returns>The builder instance with the retry strategy added.</returns>
-    public static ResiliencePipelineBuilder AddChaosFault(
-        this ResiliencePipelineBuilder builder, bool enabled, double injectionRate, Func<Exception?> faultGenerator)
+    public static ResiliencePipelineBuilder<TResult> AddChaosFault<TResult>(
+        this ResiliencePipelineBuilder<TResult> builder, bool enabled, double injectionRate, Func<Exception?> faultGenerator)
     {
         Guard.NotNull(builder);
 
@@ -54,15 +56,16 @@ internal static partial class OutcomePipelineBuilderExtensions
     /// <summary>
     /// Adds a fault chaos strategy to the builder.
     /// </summary>
+    /// <typeparam name="TResult">The type of result the retry strategy handles.</typeparam>
     /// <param name="builder">The builder instance.</param>
     /// <param name="options">The fault strategy options.</param>
     /// <returns>The builder instance with the retry strategy added.</returns>
-    public static ResiliencePipelineBuilder AddChaosFault(this ResiliencePipelineBuilder builder, FaultStrategyOptions options)
+    public static ResiliencePipelineBuilder<TResult> AddChaosFault<TResult>(this ResiliencePipelineBuilder<TResult> builder, FaultStrategyOptions options)
     {
         Guard.NotNull(builder);
         Guard.NotNull(options);
 
-        builder.AddFaultCore(options);
+        builder.AddFaultCore<TResult>(options);
         return builder;
     }
 
@@ -70,12 +73,12 @@ internal static partial class OutcomePipelineBuilderExtensions
         "Trimming",
         "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
         Justification = "All options members preserved.")]
-    private static void AddFaultCore(this ResiliencePipelineBuilder builder, FaultStrategyOptions options)
+    private static void AddFaultCore<TResult>(
+        this ResiliencePipelineBuilder<TResult> builder,
+        FaultStrategyOptions options)
     {
-        builder.AddStrategy(context =>
-            new OutcomeChaosStrategy<object>(
-                options,
-                context.Telemetry),
+        builder.AddStrategy(
+            context => new FaultChaosStrategy<TResult>(options, context.Telemetry),
             options);
     }
 }
