+++ src/Polly.Core/Simmy/Outcomes/OutcomeConstants.cs
@@ -3,6 +3,4 @@
 internal static class OutcomeConstants
 {
     public const string OnOutcomeInjectedEvent = "OnOutcomeInjected";
-
-    public const string OnFaultInjectedEvent = "OnFaultInjectedEvent";
 }
