+++ test/Polly.Core.Tests/Simmy/Fault/FaultConstantsTests.cs
@@ -1,0 +1,12 @@
+using Polly.Simmy.Fault;
+
+namespace Polly.Core.Tests.Simmy.Fault;
+
+public class FaultConstantsTests
+{
+    [Fact]
+    public void EnsureDefaults()
+    {
+        FaultConstants.OnFaultInjectedEvent.Should().Be("OnFaultInjectedEvent");
+    }
+}
