+++ test/Polly.Core.Tests/Simmy/Outcomes/OutcomeChaosStrategyTests.cs
@@ -1,5 +1,4 @@
-﻿using Polly.Simmy;
-using Polly.Simmy.Outcomes;
+﻿using Polly.Simmy.Outcomes;
 using Polly.Telemetry;
 
 namespace Polly.Core.Tests.Simmy.Outcomes;
@@ -10,22 +9,6 @@ public class OutcomeChaosStrategyTests
     private readonly List<TelemetryEventArguments<object, object>> _args = new();
 
     public OutcomeChaosStrategyTests() => _telemetry = TestUtilities.CreateResilienceTelemetry(arg => _args.Add(arg));
-
-    public static List<object[]> FaultCtorTestCases =>
-        new()
-        {
-                new object[] { null!, "Value cannot be null. (Parameter 'options')", typeof(ArgumentNullException) },
-                new object[]
-                {
-                    new FaultStrategyOptions
-                    {
-                        InjectionRate = 1,
-                        Enabled = true,
-                    },
-                    "Either Fault or FaultGenerator is required.",
-                    typeof(InvalidOperationException)
-                },
-        };
 
     public static List<object[]> ResultCtorTestCases =>
         new()
@@ -42,27 +25,6 @@ public class OutcomeChaosStrategyTests
                     typeof(InvalidOperationException)
                 },
         };
-
-    [Theory]
-    [MemberData(nameof(FaultCtorTestCases))]
-#pragma warning disable xUnit1026 // Theory methods should use all of their parameters
-    public void FaultInvalidCtor(object options, string expectedMessage, Type expectedException)
-#pragma warning restore xUnit1026 // Theory methods should use all of their parameters
-    {
-#pragma warning disable CA1031 // Do not catch general exception types
-        try
-        {
-            var _ = new OutcomeChaosStrategy<object>((FaultStrategyOptions)options, _telemetry);
-        }
-        catch (Exception ex)
-        {
-            Assert.IsType(expectedException, ex);
-#if !NET481
-            Assert.Equal(expectedMessage, ex.Message);
-#endif
-        }
-#pragma warning restore CA1031 // Do not catch general exception types
-    }
 
     [Theory]
     [MemberData(nameof(ResultCtorTestCases))]
@@ -83,174 +45,6 @@ public class OutcomeChaosStrategyTests
 #endif
         }
 #pragma warning restore CA1031 // Do not catch general exception types
-    }
-
-    [Fact]
-    public void Given_not_enabled_should_not_inject_fault()
-    {
-        var userDelegateExecuted = false;
-        var fault = new InvalidOperationException("Dummy exception");
-
-        var options = new FaultStrategyOptions
-        {
-            InjectionRate = 0.6,
-            Enabled = false,
-            Randomizer = () => 0.5,
-            Fault = fault
-        };
-
-        var sut = new ResiliencePipelineBuilder().AddChaosFault(options).Build();
-        sut.Execute(() => { userDelegateExecuted = true; });
-
-        userDelegateExecuted.Should().BeTrue();
-    }
-
-    [Fact]
-    public async Task Given_not_enabled_should_not_inject_fault_and_return_outcome()
-    {
-        var userDelegateExecuted = false;
-        var fault = new InvalidOperationException("Dummy exception");
-
-        var options = new FaultStrategyOptions
-        {
-            InjectionRate = 0.6,
-            Enabled = false,
-            Randomizer = () => 0.5,
-            Fault = fault
-        };
-
-        var sut = new ResiliencePipelineBuilder<HttpStatusCode>().AddChaosFault(options).Build();
-        var response = await sut.ExecuteAsync(async _ =>
-        {
-            userDelegateExecuted = true;
-            return await Task.FromResult(HttpStatusCode.OK);
-        });
-
-        response.Should().Be(HttpStatusCode.OK);
-        userDelegateExecuted.Should().BeTrue();
-    }
-
-    [Fact]
-    public async Task Given_enabled_and_randomly_within_threshold_should_inject_fault_instead_returning_outcome()
-    {
-        var onFaultInjected = false;
-        var userDelegateExecuted = false;
-        var exceptionMessage = "Dummy exception";
-        var fault = new InvalidOperationException(exceptionMessage);
-
-        var options = new FaultStrategyOptions
-        {
-            InjectionRate = 0.6,
-            Enabled = true,
-            Randomizer = () => 0.5,
-            Fault = fault,
-            OnFaultInjected = args =>
-            {
-                args.Context.Should().NotBeNull();
-                args.Context.CancellationToken.IsCancellationRequested.Should().BeFalse();
-                onFaultInjected = true;
-                return default;
-            }
-        };
-
-        var sut = new ResiliencePipelineBuilder<HttpStatusCode>().AddChaosFault(options).Build();
-        await sut.Invoking(s => s.ExecuteAsync(async _ =>
-        {
-            userDelegateExecuted = true;
-            return await Task.FromResult(HttpStatusCode.OK);
-        }).AsTask())
-            .Should()
-            .ThrowAsync<InvalidOperationException>()
-            .WithMessage(exceptionMessage);
-
-        userDelegateExecuted.Should().BeFalse();
-        onFaultInjected.Should().BeTrue();
-    }
-
-    [Fact]
-    public async Task Given_enabled_and_randomly_within_threshold_should_inject_fault()
-    {
-        var onFaultInjected = false;
-        var userDelegateExecuted = false;
-        var exceptionMessage = "Dummy exception";
-        var fault = new InvalidOperationException(exceptionMessage);
-
-        var options = new FaultStrategyOptions
-        {
-            InjectionRate = 0.6,
-            Enabled = true,
-            Randomizer = () => 0.5,
-            Fault = fault,
-            OnFaultInjected = args =>
-            {
-                args.Context.Should().NotBeNull();
-                args.Context.CancellationToken.IsCancellationRequested.Should().BeFalse();
-                onFaultInjected = true;
-                return default;
-            }
-        };
-
-        var sut = CreateSut<int>(options);
-        await sut.Invoking(s => s.ExecuteAsync(async _ =>
-        {
-            userDelegateExecuted = true;
-            return await Task.FromResult(200);
-        }).AsTask())
-            .Should()
-            .ThrowAsync<InvalidOperationException>()
-            .WithMessage(exceptionMessage);
-
-        userDelegateExecuted.Should().BeFalse();
-        _args.Should().HaveCount(1);
-        _args[0].Arguments.Should().BeOfType<OnFaultInjectedArguments>();
-        _args[0].Event.EventName.Should().Be(OutcomeConstants.OnFaultInjectedEvent);
-        onFaultInjected.Should().BeTrue();
-    }
-
-    [Fact]
-    public void Given_enabled_and_randomly_not_within_threshold_should_not_inject_fault()
-    {
-        var userDelegateExecuted = false;
-        var fault = new InvalidOperationException("Dummy exception");
-
-        var options = new FaultStrategyOptions
-        {
-            InjectionRate = 0.3,
-            Enabled = true,
-            Randomizer = () => 0.5,
-            Fault = fault
-        };
-
-        var sut = CreateSut<int>(options);
-        var result = sut.Execute(_ =>
-        {
-            userDelegateExecuted = true;
-            return 200;
-        });
-
-        result.Should().Be(200);
-        userDelegateExecuted.Should().BeTrue();
-    }
-
-    [Fact]
-    public void Given_enabled_and_randomly_within_threshold_should_not_inject_fault_when_exception_is_null()
-    {
-        var userDelegateExecuted = false;
-        var options = new FaultStrategyOptions
-        {
-            InjectionRate = 0.6,
-            Enabled = true,
-            Randomizer = () => 0.5,
-            FaultGenerator = (_) => new ValueTask<Exception?>(Task.FromResult<Exception?>(null))
-        };
-
-        var sut = new ResiliencePipelineBuilder().AddChaosFault(options).Build();
-        sut.Execute(_ =>
-        {
-            userDelegateExecuted = true;
-        });
-
-        userDelegateExecuted.Should().BeTrue();
     }
 
     [Fact]
@@ -392,9 +186,6 @@ public class OutcomeChaosStrategyTests
     }
 
     private ResiliencePipeline<TResult> CreateSut<TResult>(OutcomeStrategyOptions<TResult> options) =>
-        new OutcomeChaosStrategy<TResult>(options, _telemetry).AsPipeline();
-
-    private ResiliencePipeline<TResult> CreateSut<TResult>(FaultStrategyOptions options) =>
         new OutcomeChaosStrategy<TResult>(options, _telemetry).AsPipeline();
 }
