+++ test/Polly.Core.Tests/Simmy/Outcomes/OutcomeConstantsTests.cs
@@ -7,7 +7,6 @@ public class OutcomeConstantsTests
     [Fact]
     public void EnsureDefaults()
     {
-        OutcomeConstants.OnFaultInjectedEvent.Should().Be("OnFaultInjectedEvent");
         OutcomeConstants.OnOutcomeInjectedEvent.Should().Be("OnOutcomeInjected");
     }
 }
